refactor: bottom nav bar, top nav bar 경로 수정
