+++ shared/ui/index.ts
@@ -1,5 +1,4 @@
 export * from "@/shared/ui/Avatar";
-export * from "@/shared/ui/BottomNavBar";
 export * from "@/shared/ui/Button";
 export * from "@/shared/ui/ButtonGroups";
 export * from "@/shared/ui/Chip";
@@ -8,4 +7,3 @@ export * from "@/shared/ui/IconButton";
 export * from "@/shared/ui/ProfileButton";
 export * from "@/shared/ui/Switch";
 export * from "@/shared/ui/TextInput";
-export * from "@/shared/ui/TopNavBar";
